Rewrite the `registry-url-in-lock-file` RFC (#99)
* Rewrite the `registry-url-in-lock-file` RFC

The `registry-url-in-lock-file` RFC was a proposal for allowing users to more
easily switch between registries. The current problem with switching
registries is that Yarn tries to use the cached tarball URL in the `resolved`
field of the lockfile instead of the current registry when installing.

The original proposal was to remove the registry from the lockfile. This would
force Yarn to use the current registry. Later in the discussion it was also
proposed to follow the behavior of `npm@5` and just ignore the `resolved`
field.

Both of these solutions seemed to have the support of everyone involved in the
discussion, but the concern of this being a breaking change was raised. So
instead, a third proposal for an additional configuration field was what ended
up being merged.

Circumstances have changed, and a major version bump is likely in the near
future. Given that we can now make a breaking change, I am proposing that we
rewrite this RFC to more closely match its original intention. This rewrite
proposes that we solve the problem by adopting the same behavior as `npm@5`.

I have also updated the title of the RFC to better match the proposal.

* Reword question to aviod ambiguity